requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
blinker==1.9.0
click==8.3.0
colorama==0.4.6
discord.py==2.6.4
dnspython==1.16.0
Flask==3.1.2
frozenlist==1.8.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
motor==3.7.1
multidict==6.7.0
propcache==0.4.1
pymongo==4.15.3
python-dotenv==1.2.1
Werkzeug==3.1.3
yarl==1.22.0
